Lab 10
Stored procedures aren't too bad. This contains the create statements of
the database, as well as the create functions statement. I tested it on
my Postgres and everything works when you call them.

I am currently taking Discrete math, and my desire to look at anything
considered recursive is nonexistant. So unfortunately I will not be
classified a Jedi for this.