Design Project
Here is all of the sql code, the presentation, and the EDR for The
Database of Heroes. This database details Boku no Hero Academia (or My
Hero Academy) and all it's characters, quirks (powers), and the like. I
added the EDR again seperately because I think it's really pretty.

This took me a stupid amount of time, but I learned so much. While I'm
actually physically beaten down by finishing this, I feel accomplished
in designing and starting up my very own database.